backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
numpy
google-adk
google-genai
urllib3<2
